component creation and upload added

+++ abriocli/component/component.py
@@ -1,7 +1,19 @@
 import click, os ,zipfile, json
+import requests
+import datetime
+from terminaltables import AsciiTable
+
+from requests.auth import HTTPBasicAuth
+from requests_toolbelt import MultipartEncoder, MultipartEncoderMonitor
+from clint.textui.progress import Bar as ProgressBar
+
 from ..util.checker import *
 from ..conf.conf import config, get_full_path
-from ..project.project import add_component_project, remove_component_project
+from ..project.project import add_component_project, remove_component_project, load_project_config,\
+    load_component_config,write_component_config
+from ..util.checker import ensure_component_exists
+from ..conf.conf import config,errors
+
 
 @click.option('--version', prompt="Enter component version", default="0.0.1" )
 @click.option('--public', is_flag=True, prompt='Do you want to mark this component as public?' )
@@ -20,36 +32,91 @@ def init(name, public, version):
         click.secho("\nDirecotry with name <{0}> already exists.\n".format(name), fg="red", bold=True)
         return
 
+    click.secho("\nConnection to sever..." , bold=True, fg="yellow")
+    project_config = load_project_config()
+    email = project_config['email']
+    pwd = project_config['password']
+    name = name
+    is_private = public
+    response = requests.post(
+        config['server']['host']+'component/create',
+        auth=HTTPBasicAuth(email, pwd),
+        json={'isPrivate': is_private, 'name': name}
+    )
 
-    os.mkdir(name)
-    zfile = zipfile.ZipFile(os.path.join(get_full_path('data', config['sdk_file'])))
-    zfile.extractall(name)
+    if response.status_code == 201 :
+        pkey = response.json()['token']
+        os.mkdir(name)
+        zip_file = zipfile.ZipFile(os.path.join(get_full_path('data', config['sdk_file'])))
+        zip_file.extractall(name)
 
-    # make a request to server and create this component
-    component_config = {
-        'pkey' : '',
-        'public' : public,
-        'version': version,
-        'name' : name,
-        'last_compiled' : ''
-    }
+        component_config = {
+            'pkey': pkey,
+            'public': public,
+            'version': version,
+            'name': name,
+            'last_uploaded': ''
+        }
 
-    with open(os.path.join(name, (name+'.json')), 'w') as config_file :
-        config_file.write(json.dumps(component_config, indent=4, separators=(',', ': ')))
+        # with open(os.path.join(name, (name+'.json')), 'w') as config_file :
+        #     config_file.write(json.dumps(component_config, indent=4, separators=(',', ': ')))
 
-    add_component_project(name)
+        add_component_project(component_config)
 
-    click.secho("\nComponent <{0}> created.\n".format(name), bold=True, fg='green')
+        click.secho("\nComponent <{0}> created.\n".format(name), bold=True, fg='green')
 
-
+    else :
+        click.secho(errors["UNKNOWN_NETWORK"],bold=True, fg="red")
 
 
 @click.command()
-def upload() :
+@click.argument('name')
+def upload(name) :
     '''
     Upload Abrio component to server.
     '''
-    pass
+
+    if not ensure_abrio_root():
+        click.secho('\nAbrio Root Directory Not Detected.\n', fg="red", bold=True)
+        return
+
+    if not ensure_component_exists(name):
+        click.secho("\nComponent <{0}> does not exist.\n".format(name), bold=True, fg="red")
+
+    build_dir = '/sample/build/libs/'
+
+    os.system('cd {0} && gradle jar && cd ..'.format(name))
+    jar_dir = name + build_dir + name + '.jar'
+    os.rename(name + build_dir + 'sample.jar',jar_dir)
+
+    encoder = create_upload(jar_dir)
+    callback = create_callback(encoder)
+    monitor = MultipartEncoderMonitor(encoder, callback)
+
+    component_config = load_component_config(name)
+    component_config['last_uploaded'] = str(datetime.datetime.now())
+    write_component_config(name, component_config)
+
+    headers = {
+        'Content-Type': monitor.content_type,
+        'private key': component_config['pkey'],
+        'version' : component_config['version']
+    }
+
+    upload_response = requests.post(
+        config['server']['host'] + "component/upload",
+        data=monitor,
+        # auth=HTTPBasicAuth(email, pwd),
+        headers=headers)
+
+    if upload_response.status_code == 200 :
+        click.secho('\n\n\nComponent uploaded\n', bold=True, fg="green")
+
+    else :
+        click.secho(errors["UNKNOWN_NETWORK"], bold=True, fg="red")
+
+
+
 
 
 @click.option('--sure', prompt="Are you sure you want to delete this component", is_flag=True)
@@ -59,14 +126,55 @@ def rm(name, sure) :
     '''
     Delete Abrio Component.
     '''
+
+    if not ensure_abrio_root():
+        click.secho('\nAbrio Root Directory Not Detected.\n', fg="red", bold=True)
+        return
+
     if sure :
-        os.system('rm -Rf {0}'.format(name))
-        remove_component_project(name)
-        click.secho("\nComponent <{0}> deleted.\n".format(name), bold=True, fg="yellow")
+        if ensure_component_exists(name) :
+            os.system('rm -Rf {0}'.format(name))
+            remove_component_project(name)
+            # todo delete from server
+            click.secho("\nComponent <{0}> deleted.\n".format(name), bold=True, fg="yellow")
+
+        else :
+            click.secho("\nComponent <{0}> does not exist.\n".format(name), bold=True, fg="red")
 
 @click.command()
 def ls() :
     '''
     List Available Abrio components
     '''
-    pass
+
+    if not ensure_abrio_root():
+        click.secho('\nAbrio Root Directory Not Detected.\n', fg="red", bold=True)
+        return
+
+    config = load_project_config()
+    component_table = [
+        ['Component Name', 'Version', 'Public' , "Last Upload"]
+    ] + [
+            [
+                component['name'],
+                component['version'],
+                str(component['public']),
+                component['last_uploaded']
+            ] for component in config['components']
+        ]
+
+    table = AsciiTable(component_table)
+    click.echo(table.table)
+
+def create_callback(encoder):
+    encoder_len = encoder.len
+
+    bar = ProgressBar(expected_size=encoder_len, filled_char='=')
+
+    def callback(monitor):
+        bar.show(monitor.bytes_read)
+    return callback
+
+def create_upload(file_path):
+    file_name = file_path.split("/")[-1]
+    return MultipartEncoder({'files':(file_name,open(file_path, 'rb'))})
